back/hostel_app/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import (
    register_view, login_view, logout_view, current_user_view,
    admin_stats_view, UserViewSet, RoomViewSet, BedViewSet,
    ResidentViewSet, AuditLogViewSet
)

router = DefaultRouter()
router.register(r'users', UserViewSet, basename='user')
router.register(r'rooms', RoomViewSet, basename='room')
router.register(r'beds', BedViewSet, basename='bed')
router.register(r'residents', ResidentViewSet, basename='resident')
router.register(r'audit-logs', AuditLogViewSet, basename='audit-log')

urlpatterns = [
    # Authentication
    path('auth/register/', register_view, name='register'),
    path('auth/login/', login_view, name='login'),
    path('auth/logout/', logout_view, name='logout'),
    path('auth/me/', current_user_view, name='current-user'),
    
    # Admin
    path('admin/stats/', admin_stats_view, name='admin-stats'),
    
    # Router URLs
    path('', include(router.urls)),
]
